add an idle timeout for the worker server
this adds -i to the server command, as well as -i and -n to the load command,
which will be forwarded to the server process

Summary:
The workers are fairly aggressive about polling for work, and for one
off loads it'll be better to just shut down the server when it's done.

Test Plan:
Performing multiple loads of different datasets, as well as starting
the server independently with different combinations of flags to ensure neither
the server nor load commands broke.

Reviewers: carl, hurshal, tyler, john

Subscribers: engineering-list

Differential Revision: https://grizzly.memsql.com/D16034